Add NEF emulator config values
